feat: add service to create a card
This service sees if the API key points to a existing company.
Also sees if the employee is employed by the company.
Checks if the employee already have a card of the same type.
Generates card information for this employee and register the card for them.
Returns the generated information with encrypted CVV.