feat: create new route POST /cards
Create a new route to create cards.
It uses the middleware to validate header and body, and a controller to generate the card.
Returns 201 with the card information generated if success.
Returns 403 when the employee is from another company or already have a card of same type.
Returns 404 when API key does not match a registered company, or the employee id doesn't exists.
Returns 500 if anything breaks internally.